db/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
alembic
annotated-types
anyio
appdirs
asgiref
asttokens
attrs
auth0-python
Authlib
backoff
bcrypt
beautifulsoup4
blinker
build
cachetools
certifi
cffi
charset-normalizer
chroma-hnswlib
chromadb
click
cohere
coloredlogs
crewai
crewai-tools
cryptography
dataclasses-json
decorator
Deprecated
deprecation
distro
docker
docstring_parser
docx2txt
durationpy
embedchain
et_xmlfile
executing
fastapi
fastavro
filelock
flatbuffers
frozenlist
fsspec
google-api-core
google-auth
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-core
google-cloud-resource-manager
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
gptcache
greenlet
grpc-google-iam-v1
grpcio
grpcio-health-checking
grpcio-status
grpcio-tools
h11
h2
hpack
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
hyperframe
idna
ijson
importlib_metadata
importlib_resources
iniconfig
instructor
ipython
jedi
Jinja2
jiter
json_repair
jsonpatch
jsonpickle
jsonpointer
jsonref
jsonschema
jsonschema-specifications
kubernetes
lancedb
langchain
langchain-cohere
langchain-community
langchain-core
langchain-experimental
langchain-openai
langchain-text-splitters
langsmith
linkup-sdk
litellm
Mako
markdown-it-py
MarkupSafe
marshmallow
matplotlib-inline
mdurl
mem0ai
mmh3
monotonic
mpmath
multidict
mypy-extensions
networkx
nodeenv
numpy
oauthlib
onnxruntime
openai
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
outcome
overrides
packaging
pandas
parameterized
parso
pdfminer.six
pdfplumber
pexpect
pillow
pluggy
portalocker
posthog
prompt_toolkit
propcache
proto-plus
protobuf
ptyprocess
pure_eval
pyarrow
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic-settings
pydantic_core
Pygments
PyJWT
pylance
pypdf
pypdfium2
PyPika
pyproject_hooks
pyright
pysbd
PySocks
pytest
python-dateutil
python-dotenv
pytube
pytz
pyvis
PyYAML
qdrant-client
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
schema
scrapegraph_py
selenium
serpapi
setuptools
shapely
shellingham
six
sniffio
sortedcontainers
soupsieve
spider-client
SQLAlchemy
stack-data
starlette
sympy
tabulate
tenacity
tiktoken
tokenizers
tomli
tomli_w
tqdm
traitlets
trio
trio-websocket
typer
types-requests
typing-inspect
typing_extensions
tzdata
urllib3
uv
uvicorn
uvloop
validators
watchfiles
wcwidth
weaviate-client
websocket-client
websockets
wrapt
wsproto
yarl
zipp
